Don't automatically meter timers.
The problem with doing this in the Riemann layer is that we don't record the
initial "this thing happened" until we get the timing data, which is *after* the
thing has already been timed and finished.

Imagine something takes 10 seconds to run.  If it starts at 10:55:00 the old
method would record the "foo happened" event at 10:55:10 instead.  For many
things this won't matter, but for some it will, and it's especially important
when trying to determine cause/effect.

The solution is to stop doing it automatically in the Riemann layer and have the
clients manually send a Meter metric when something they're timing starts, then
send the time data later whenever they get around to it.

Looking at the timing percentiles in graphite is still going to be shifted into
the future based on when they end, but at least the *rate* metrics for Timers
will be accurate.